Mostrar foto

+++ app/Http/Controllers/EmpleadoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Empleado;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class EmpleadoController extends Controller
 {
@@ -46,7 +47,7 @@ class EmpleadoController extends Controller
         Empleado::insert($datosempleado);
 
         //return response()->json($datosempleado);
-        return redirect('empleado');
+        return redirect('empleado')->with('mensaje','Empleado agregado con éxito');
     }
 
     /**
@@ -66,9 +67,10 @@ class EmpleadoController extends Controller
      * @param  \App\Models\Empleado  $empleado
      * @return \Illuminate\Http\Response
      */
-    public function edit(Empleado $empleado)
+    public function edit($id)
     {
-        
+        $empleado=Empleado::findOrFail($id);
+        return view('empleado.edit',compact('empleado'));
     }
 
     /**
@@ -78,9 +80,21 @@ class EmpleadoController extends Controller
      * @param  \App\Models\Empleado  $empleado
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Empleado $empleado)
+    public function update(Request $request, $id)
     {
-        //
+        $datosempleado=request()->except('_token','_method');
+
+        if($request->hasFile('Foto')){
+            $empleado=Empleado::findOrFail($id);
+            Storage::delete('public/'.$empleado->Foto);
+            $datosempleado['Foto']=$request->file('Foto')->store('uploads','public');
+        }
+
+        Empleado::where('id','=',$id)->update($datosempleado);
+
+
+        $empleado=Empleado::findOrFail($id);
+        return view('empleado.edit',compact('empleado'));
     }
 
     /**
@@ -91,7 +105,12 @@ class EmpleadoController extends Controller
      */
     public function destroy($id)
     {
-        Empleado::destroy($id);
-        return redirect('empleado');
+        $empleado=Empleado::findOrFail($id);
+        
+        if(Storage::delete('public/'.$empleado->Foto)){
+            Empleado::destroy($id);
+        }
+
+        return redirect('empleado')->with('mensaje','Empleado eliminado');;
     }
 }
